// packages/integrations/zendesk/generated/help-center/types.gen.ts
export type ArticleObject = {
    /**
     * The id of the user who wrote the article (set to the user who made the request on create by default)
     */
    author_id?: number;
    /**
     * HTML body of the article. Unsafe tags and attributes may be removed before display. For a list of safe tags and attributes, see [Allowing unsafe HTML in Help Center articles](https://support.zendesk.com/hc/en-us/articles/115015895948) in Zendesk help
     */
    body?: string;
    /**
     * True if comments are disabled; false otherwise
     */
    comments_disabled?: boolean;
    /**
     * The list of content tags attached to the article
     */
    content_tag_ids?: Array<string>;
    /**
     * The time the article was created
     */
    readonly created_at?: string;
    /**
     * True if the translation for the current locale is a draft; false otherwise. false by default. Can be set when creating but not when updating. For updating, see Translations
     */
    readonly draft?: boolean;
    /**
     * The time the article was last edited in its displayed locale
     */
    readonly edited_at?: string;
    /**
     * The url of the article in Help Center
     */
    readonly html_url?: string;
    /**
     * Automatically assigned when the article is created
     */
    readonly id?: number;
    /**
     * An array of label names associated with this article. By default no label names are used. Only available on certain plans
     */
    label_names?: Array<string>;
    /**
     * The locale that the article is being displayed in
     */
    locale: string;
    /**
     * Deprecated. Always false because the source translation is always the most up-to-date translation
     */
    readonly outdated?: boolean;
    /**
     * Locales in which the article was marked as outdated
     */
    readonly outdated_locales?: Array<string>;
    /**
     * The id of the permission group which defines who can edit and publish this article
     */
    permission_group_id: number;
    /**
     * The position of this article in the article list. 0 by default
     */
    position?: number;
    /**
     * True if this article is promoted; false otherwise. false by default
     */
    promoted?: boolean;
    /**
     * The id of the section to which this article belongs
     */
    section_id?: number;
    /**
     * The source (default) locale of the article
     */
    readonly source_locale?: string;
    /**
     * The title of the article
     */
    title: string;
    /**
     * The time the article was last updated
     */
    readonly updated_at?: string;
    /**
     * The API url of the article
     */
    readonly url?: string;
    /**
     * The id of the user segment which defines who can see this article. Set to null to make it accessible to everyone. Either user_segment_id or user_segment_ids must be specified
     */
    user_segment_id?: number | null;
    /**
     * List of user segment ids which define who can view this article. Set to an empty list to make it accessible to everyone. For Enterprise plans only this may contain more than one user_segment_id. Either user_segment_id or user_segment_ids must be specified
     */
    user_segment_ids?: Array<number>;
    /**
     * The total number of upvotes and downvotes
     */
    readonly vote_count?: number;
    /**
     * The sum of upvotes (+1) and downvotes (-1), which may be positive or negative
     */
    readonly vote_sum?: number;
};
// ...
export type CategoryObject = {
    /**
     * The time at which the category was created
     */
    readonly created_at?: string;
    /**
     * The description of the category
     */
    description?: string;
    /**
     * The url of this category in Help Center
     */
    readonly html_url?: string;
    /**
     * Automatically assigned when creating categories
     */
    readonly id: number;
    /**
     * The locale where the category is displayed
     */
    locale: string;
    /**
     * The name of the category
     */
    name: string;
    /**
     * Whether the category is out of date
     */
    readonly outdated?: boolean;
    /**
     * The position of this category relative to other categories
     */
    position?: number;
    /**
     * The source (default) locale of the category
     */
    readonly source_locale?: string;
    /**
     * The time at which the category was last updated
     */
    readonly updated_at?: string;
    /**
     * The API url of this category
     */
    readonly url?: string;
};
// ...
export type SectionObject = {
    /**
     * The id of the category to which this section belongs
     */
    category_id?: number;
    /**
     * The time at which the section was created
     */
    readonly created_at?: string;
    /**
     * The description of the section
     */
    description?: string;
    /**
     * The url of this section in HC
     */
    readonly html_url?: string;
    /**
     * Automatically assigned when creating subscriptions
     */
    readonly id?: number;
    /**
     * The locale in which the section is displayed
     */
    locale: string;
    /**
     * The name of the section
     */
    name: string;
    /**
     * Whether the section is out of date
     */
    readonly outdated?: boolean;
    /**
     * The id of the section to which this section belongs. Only writable for Guide Enterprise customers
     */
    parent_section_id?: number | null;
    /**
     * The position of this section in the section list. Used when sorting is set to ´manual´. By default the section is added to the end of the list
     */
    position?: number;
    /**
     * The source (default) locale of the section
     */
    readonly source_locale?: string;
    /**
     * The theme template name used to display this section in Help Center.
     */
    theme_template?: string;
    /**
     * The time at which the section was last updated
     */
    readonly updated_at?: string;
    /**
     * The API url of this section
     */
    readonly url?: string;
};